# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.BatchCreateRedemptionCodesRequest:
    properties:
      bandwidthId:
        type: string
      count:
        maximum: 100
        minimum: 1
        type: integer
      cpuId:
        type: string
      diskId:
        type: string
      imageId:
        type: integer
      instanceType:
        enum:
        - container
        - vm
        type: string
      memoryId:
        type: string
      providerId:
        type: integer
      remark:
        type: string
    required:
    - bandwidthId
    - count
    - cpuId
    - diskId
    - imageId
    - instanceType
    - memoryId
    - providerId
    type: object
...
  admin.BatchDeleteRedemptionCodesRequest:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  admin.ExportRedemptionCodesRequest:
    properties:
      fields:
        description: 要导出的字段列表，为空则导出所有字段
        items:
          type: string
        type: array
      ids:
        description: 为空则导出所有
        items:
          type: integer
        type: array
      lang:
        description: '语言: zh-CN, en-US'
        type: string
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  config.UnifiedConfigRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      scope:
        description: public, user, admin
        type: string
    required:
    - config
    - scope
    type: object
...
host: 0.0.0.0:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 一键虚拟化管理平台API接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OneClickVirt API
  version: "1.0"
paths:
...
  /admin/redemption-codes:
    get:
      consumes:
      - application/json
      description: 管理员获取兑换码列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 兑换码关键字
        in: query
        name: code
        type: string
      - description: 状态筛选
        in: query
        name: status
        type: string
      - description: 节点ID筛选
        in: query
        name: providerId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取兑换码列表
      tags:
      - 兑换码管理
  /admin/redemption-codes/batch-create:
    post:
      consumes:
      - application/json
      description: 管理员批量创建兑换码，触发后台实例创建任务
      parameters:
      - description: 批量创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.BatchCreateRedemptionCodesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 批量创建兑换码
      tags:
      - 兑换码管理
  /admin/redemption-codes/batch-delete:
    post:
      consumes:
      - application/json
      description: 管理员批量删除兑换码（硬删除）及关联实例
      parameters:
      - description: 批量删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.BatchDeleteRedemptionCodesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 批量删除兑换码
      tags:
      - 兑换码管理
  /admin/redemption-codes/export:
    post:
      consumes:
      - application/json
      description: 管理员导出指定兑换码的字符串列表
      parameters:
      - description: 导出请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.ExportRedemptionCodesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 导出成功
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 导出兑换码
      tags:
      - 兑换码管理
...
  /config:
    get:
      consumes:
      - application/json
      description: 根据用户权限返回相应的配置信息
      parameters:
      - default: user
        description: 配置范围
        enum:
        - public
        - user
        - admin
        in: query
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data: {}
              type: object
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取系统配置
      tags:
      - 配置管理
    put:
      consumes:
      - application/json
      description: 根据用户权限更新相应的配置信息
      parameters:
      - description: 配置更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/config.UnifiedConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新系统配置
      tags:
      - 配置管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"